Extracting reading of the conf and the job to ConfigurationHelper.
